requirements.txt
pandas
requests
python-dotenv
xgboost
matplotlib